# PE/lab04/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "API приложения для хранения файлов",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8099"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create User",
                "operationId": "create_user_api_v1_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "detail": "User not found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "detail": "Invalid user data"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "detail": "Internal error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{user_id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get User",
                "operationId": "get_user_api_v1_users__user_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Unauthorized",
                        "description": "Unauthorized"
                    },
                    "404": {
                        "detail": "User not found",
                        "description": "Not Found"
                    },
                    "403": {
                        "detail": "Access denied",
                        "description": "Forbidden"
                    },
                    "500": {
                        "detail": "Internal error",
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get Users",
                "operationId": "get_users_api_v1_users_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "Login"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "First Name"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "Last Name"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    },
                                    "title": "Response Get Users Api V1 Users Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Unauthorized",
                        "description": "Unauthorized"
                    },
                    "403": {
                        "detail": "Access denied",
                        "description": "Forbidden"
                    },
                    "500": {
                        "detail": "Internal error",
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "login_api_v1_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokensResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "detail": "Login unable"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "detail": "Internal error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Refresh",
                "operationId": "refresh_api_v1_refresh_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshToken"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokensResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "detail": "Refresh token has expired"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "detail": "Internal error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{user_id}/files": {
            "post": {
                "tags": [
                    "core"
                ],
                "summary": "Create File In Root Folder",
                "operationId": "create_file_in_root_folder_api_v1_users__user_id__files_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_create_file_in_root_folder_api_v1_users__user_id__files_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "detail": "Object not found",
                        "description": "Not Found"
                    },
                    "409": {
                        "detail": "Object already exists",
                        "description": "Conflict"
                    },
                    "401": {
                        "detail": "Unauthorized",
                        "description": "Unauthorized"
                    },
                    "400": {
                        "detail": "Object name is invalid",
                        "description": "Bad Request"
                    },
                    "403": {
                        "detail": "Access denied",
                        "description": "Forbidden"
                    },
                    "500": {
                        "detail": "Internal error",
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "LoginRequest": {
                "properties": {
                    "login": {
                        "type": "string",
                        "title": "Login"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "login",
                    "password"
                ],
                "title": "LoginRequest"
            },
...
            "UserCreate": {
                "properties": {
                    "login": {
                        "type": "string",
                        "title": "Login"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "login",
                    "first_name",
                    "last_name",
                    "email",
                    "password"
                ],
                "title": "UserCreate"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}